pyproject.toml
[project]
name = "win-bc-controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bumble>=0.0.220",
    "inputs>=0.5",
    "pygame-ce>=2.5.6",
    "websockets>=15.0.1",
]
